Migrate `tests/unit/modules/test_azurearm_dns.py` to PyTest
